Working on making serach functionality work
