+++ Application/Routes.py
@@ -1,11 +1,19 @@
-from flask import render_template
+from flask import render_template, redirect
 from Application import app
 from flask import request, session
 from Application.Database_Funcs.User import verify_login, create_user
+from Application.Database_Funcs.Place import get_places, add_place
 from Application.Models import User
 
-@app.route("/")
+@app.route("/", methods=["GET", "POST"])
 def home():
+    if (request.method == "POST"):
+        print("PARAMS: ", request.form["checkin"], request.form["checkout"], request.form["num_guests"], request.form["miles_campus"])
+
+        # Fetch all the places from the database
+        places_list = get_places()
+        return render_template("Search/Places.html", places=places_list)
+
     return render_template("Home/Home.html")
 
 #TODO replace username with the actual user's username?
@@ -57,3 +65,26 @@ def createUser():
 def redirect_logged_in(user: User):
     print("User: ", user)
     return render_template("User/Home.html", user=user)
+
+# Route to add a new place TODO, have it only available if a user is logged in
+@app.route("/add_place", methods=["GET", "POST"])
+def create_place():
+    error = None
+    if request.method == "POST":
+        # Retrieve form data
+        place_name = request.form["place_name"]
+        place_type = request.form["type"]
+        price = request.form["price"]
+        amenities = request.form["amenities"]
+        rating = request.form["rating"]
+        campus_distance = request.form["campus_distance"]
+
+        # Add the place to the database
+        did_add_place = add_place(place_name, place_type, price, amenities, rating, campus_distance)
+
+        if did_add_place:
+            return redirect("/places")
+        else:
+            error = "Failed to add the place. Please try again."
+
+    return render_template("Places/AddPlace.html", error=error)
