+++ Application/Templates/Home/Home.html
@@ -76,42 +76,43 @@
             <h1 class="text-4xl sm:text-5xl font-bold mb-4">Welcome to Swamp Stays</h1>
             <p class="text-lg sm:text-xl mb-6 text-white">Find and book your perfect stay close to campus with ease!</p>
 
-            
-            <div class="dropdown">
-                <button onclick="toggleDropdown()" class="dropbtn">Number of guests</button>
-                <div id="myDropdown" class="dropdown-content">
-                    <input type="text" placeholder="Search.." id="myInput" onkeyup="filterFunction()">
-                    <a href="#">1</a>
-                    <a href="#">2</a>
-                    <a href="#">3</a>
-                    <a href="#">4</a>
-                    <a href="#">5</a>
-                    <a href="#">6</a>
-                    <a href="#">7</a>
-                    <a href="#">8</a>
+            <form method="POST">
+                <div class="dropdown">
+                    <button type="button" onclick="toggleDropdown()" class="dropbtn">Number of guests</button>
+                    <div id="myDropdown" class="dropdown-content">
+                        <input type="text" placeholder="Search.." id="myInput" name="num_guests" onkeyup="filterFunction()">
+                        <a href="#">1</a>
+                        <a href="#">2</a>
+                        <a href="#">3</a>
+                        <a href="#">4</a>
+                        <a href="#">5</a>
+                        <a href="#">6</a>
+                        <a href="#">7</a>
+                        <a href="#">8</a>
 
+                    </div>
                 </div>
-            </div>
 
-            <label for ="checkin" class="text-white">Check in</label>
-            <input type="date" id="checkin" name="checkin">
-            <label for ="checkout" class="text-white">Check out</label>
-            <input type="date" id="checkout" name="checkout">
+                <label for ="checkin" class="text-white">Check in</label>
+                <input type="date" id="checkin" name="checkin">
+                <label for ="checkout" class="text-white">Check out</label>
+                <input type="date" id="checkout" name="checkout">
 
-            <div class="dropdown">
-                <button onclick="toggleDropdown()" class="dropbtn">Distance from campus</button>
-                <div id="myDropdown" class="dropdown-content">
-                    <input type="text" placeholder="Search.." id="myInput" onkeyup="filterFunction()">
-                    <a href="#">under 5 min</a>
-                    <a href="#">5 min</a>
-                    <a href="#">10 min</a>
-                    <a href="#">15 min</a>
+                <div class="dropdown">
+                    <button type="button" onclick="toggleDropdown()" class="dropbtn">Distance from campus</button>
+                    <div id="myDropdown" class="dropdown-content">
+                        <input type="text" placeholder="Search.." id="myInput" name="miles_campus" onkeyup="filterFunction()">
+                        <a href="#">under 5 min</a>
+                        <a href="#">5 min</a>
+                        <a href="#">10 min</a>
+                        <a href="#">15 min</a>
+                    </div>
                 </div>
-            </div>
 
-            <button class="flex justify-center items-center bg-white text-blue-800 font-semibold px-6 py-3 rounded-full shadow-md hover:bg-blue-700 hover:text-white transition">
-                Start Booking
-            </button>
+                <button type="submit"  class="flex justify-center items-center bg-white text-blue-800 font-semibold px-6 py-3 rounded-full shadow-md hover:bg-blue-700 hover:text-white transition">
+                    Start Booking
+                </button>
+            </form>
         </div>
     </div>
