+++ Application/Templates/Search/Places.html
@@ -11,7 +11,6 @@
         <li>{{ place.place_name }} - {{ place.type }} - ${{ place.price }} per night</li>
         {% endfor %}
     </ul>
-    <a href="/add_place">Add a New Place</a>
 </body>
 </html>
